Module minerva_data::db

This module contains functions and structures related to handling the PostgreSQL database, specially regarding connections and database creation.

Modules

create 🔒

Functions

build_database_string

Generates a string to connect to PostgreSQL RDBMS, given a tenant name and the database server endpoint (e.g. localhost:5432). This function assumes a user postgres with a password postgres.

create_database

Runs a database creation query, if said database does not exist. The database in question shall have the same name of a tenant. This function also expects the endpoint to a database server (e.g. localhost:5432).

make_connection_pool

Creates a connection pool to the PostgreSQL RDBMS, given a maximum number of connections, and panics if the connections could not be established.

make_single_connection

Creates a single connection to the PostgreSQL RDBMS, and panics if the connection could not be established.

try_make_single_connection

Attempts to generate a single connection to the PostgreSQL RDBMS, without error checks. This could be used to perform connections and evaluate if they could actually be established.

Type Definitions

DBConnection

Type representing a Diesel connection to a database.

DBPool

Type representing a Diesel connection pool to a database.