Struct minerva_data::syslog::NewLog

pub struct NewLog {
    pub service: String,
    pub requestor: String,
    pub entity: String,
    pub operation: OpType,
    pub datetime: DateTime<Utc>,
    pub description: Option<String>,
}

DTO for adding a new entry on the syslog table.

Fields
service: String

Name of the service performing the operation.

requestor: String

Username of whom is requesting the operation.

entity: String

Name of the entity being manipulated, usually the same as a table.

operation: OpType

Operation being performed.

datetime: DateTime<Utc>

Date and time of the operation being performed.

description: Option<String>

Optional description of the operation being performed.